(#34) Rename away from Blueprint along the js<>C++ border
