+++ packages/react-juce/template/src/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Blueprint from 'react-juce';
+import ReactJUCE from 'react-juce';
 import App from './App';
 
-Blueprint.render(<App />, Blueprint.getRootContainer());
+ReactJUCE.render(<App />, ReactJUCE.getRootContainer());
